refactor: extract creation of task manager to factory
