+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/tasks/BackgroundTaskManager.java
@@ -7,46 +7,22 @@
  */
 package io.camunda.exporter.tasks;
 
-import static io.camunda.zeebe.protocol.Protocol.START_PARTITION_ID;
-
-import io.camunda.exporter.ExporterResourceProvider;
-import io.camunda.exporter.config.ConnectionTypes;
-import io.camunda.exporter.config.ExporterConfiguration;
-import io.camunda.exporter.config.ExporterConfiguration.ArchiverConfiguration;
-import io.camunda.exporter.metrics.CamundaExporterMetrics;
-import io.camunda.exporter.tasks.archiver.ApplyRolloverPeriodJob;
 import io.camunda.exporter.tasks.archiver.ArchiverRepository;
-import io.camunda.exporter.tasks.archiver.BatchOperationArchiverJob;
-import io.camunda.exporter.tasks.archiver.ElasticsearchRepository;
-import io.camunda.exporter.tasks.archiver.OpenSearchRepository;
-import io.camunda.exporter.tasks.archiver.ProcessInstancesArchiverJob;
-import io.camunda.search.connect.es.ElasticsearchConnector;
-import io.camunda.search.connect.os.OpensearchConnector;
-import io.camunda.webapps.schema.descriptors.operate.ProcessInstanceDependant;
-import io.camunda.webapps.schema.descriptors.operate.template.BatchOperationTemplate;
-import io.camunda.webapps.schema.descriptors.operate.template.ListViewTemplate;
-import io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate;
 import io.camunda.zeebe.util.CloseableSilently;
 import io.camunda.zeebe.util.VisibleForTesting;
-import io.camunda.zeebe.util.error.FatalErrorHandler;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.concurrent.Executor;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
-import java.util.concurrent.ThreadFactory;
 import javax.annotation.WillCloseWhenClosed;
 import org.agrona.CloseHelper;
 import org.slf4j.Logger;
 
 public final class BackgroundTaskManager implements CloseableSilently {
-
   private final int partitionId;
   private final ArchiverRepository repository;
   private final Logger logger;
   private final ScheduledThreadPoolExecutor executor;
   private final List<Runnable> tasks;
-  private final ArchiverConfiguration config;
 
   private int submittedTasks = 0;
 
@@ -56,14 +32,12 @@ public final class BackgroundTaskManager implements CloseableSilently {
       final @WillCloseWhenClosed ArchiverRepository repository,
       final Logger logger,
       final @WillCloseWhenClosed ScheduledThreadPoolExecutor executor,
-      final List<Runnable> tasks,
-      final ArchiverConfiguration config) {
+      final List<Runnable> tasks) {
     this.partitionId = partitionId;
     this.repository = Objects.requireNonNull(repository, "must specify a repository");
     this.logger = Objects.requireNonNull(logger, "must specify a logger");
     this.executor = Objects.requireNonNull(executor, "must specify an executor");
     this.tasks = Objects.requireNonNull(tasks, "must specify tasks");
-    this.config = Objects.requireNonNull(config, "must specify a config");
   }
 
   @Override
@@ -95,170 +69,9 @@ public final class BackgroundTaskManager implements CloseableSilently {
         unsubmittedTasks,
         submittedTasks,
         tasks.size());
-
     for (; submittedTasks < tasks.size(); submittedTasks++) {
       final var task = tasks.get(submittedTasks);
       executor.submit(task);
-    }
-  }
-
-  public static BackgroundTaskManager create(
-      final int partitionId,
-      final String exporterId,
-      final ExporterConfiguration config,
-      final ExporterResourceProvider resourceProvider,
-      final CamundaExporterMetrics metrics,
-      final Logger logger) {
-    final var threadFactory =
-        Thread.ofPlatform()
-            .name("exporter-" + exporterId + "-p" + partitionId + "-tasks-", 0)
-            .uncaughtExceptionHandler(FatalErrorHandler.uncaughtExceptionHandler(logger))
-            .factory();
-    final var executor = defaultExecutor(threadFactory);
-    final var repository =
-        createRepository(config, resourceProvider, partitionId, executor, metrics, logger);
-    final List<Runnable> tasks = new ArrayList<>();
-    int threadCount = 1;
-
-    tasks.add(
-        createProcessInstanceTask(
-            metrics, logger, resourceProvider, repository, executor, config.getArchiver()));
-    if (partitionId == START_PARTITION_ID) {
-      threadCount = 2;
-      tasks.add(
-          createBatchOperationTask(
-              metrics, logger, resourceProvider, repository, executor, config.getArchiver()));
-      tasks.add(createApplyRolloverPeriodTask(metrics, logger, repository));
-    }
-
-    executor.setCorePoolSize(threadCount);
-    return new BackgroundTaskManager(
-        partitionId, repository, logger, executor, tasks, config.getArchiver());
-  }
-
-  private static ReschedulingTask createProcessInstanceTask(
-      final CamundaExporterMetrics metrics,
-      final Logger logger,
-      final ExporterResourceProvider resourceProvider,
-      final ArchiverRepository repository,
-      final ScheduledThreadPoolExecutor executor,
-      final ArchiverConfiguration config) {
-    final var dependantTemplates = new ArrayList<ProcessInstanceDependant>();
-    resourceProvider.getIndexTemplateDescriptors().stream()
-        .filter(ProcessInstanceDependant.class::isInstance)
-        .map(ProcessInstanceDependant.class::cast)
-        .forEach(dependantTemplates::add);
-
-    // add a special case just for TaskTemplate, which has 2 kinds of documents in the same
-    // index
-    final var taskTemplate = resourceProvider.getIndexTemplateDescriptor(TaskTemplate.class);
-    dependantTemplates.add(
-        new ProcessInstanceDependantAdapter(taskTemplate.getFullQualifiedName(), TaskTemplate.ID));
-
-    return new ReschedulingTask(
-        new ProcessInstancesArchiverJob(
-            repository,
-            resourceProvider.getIndexTemplateDescriptor(ListViewTemplate.class),
-            dependantTemplates,
-            metrics,
-            logger,
-            executor),
-        config.getRolloverBatchSize(),
-        config.getDelayBetweenRuns(),
-        executor,
-        logger);
-  }
-
-  private static ApplyRolloverPeriodJob createApplyRolloverPeriodTask(
-      final CamundaExporterMetrics metrics,
-      final Logger logger,
-      final ArchiverRepository repository) {
-
-    return new ApplyRolloverPeriodJob(repository, metrics, logger);
-  }
-
-  private static ReschedulingTask createBatchOperationTask(
-      final CamundaExporterMetrics metrics,
-      final Logger logger,
-      final ExporterResourceProvider resourceProvider,
-      final ArchiverRepository repository,
-      final ScheduledThreadPoolExecutor executor,
-      final ArchiverConfiguration config) {
-
-    return new ReschedulingTask(
-        new BatchOperationArchiverJob(
-            repository,
-            resourceProvider.getIndexTemplateDescriptor(BatchOperationTemplate.class),
-            metrics,
-            logger,
-            executor),
-        config.getRolloverBatchSize(),
-        config.getDelayBetweenRuns(),
-        executor,
-        logger);
-  }
-
-  private static ScheduledThreadPoolExecutor defaultExecutor(final ThreadFactory threadFactory) {
-    final var executor = new ScheduledThreadPoolExecutor(0, threadFactory);
-    executor.setContinueExistingPeriodicTasksAfterShutdownPolicy(false);
-    executor.setExecuteExistingDelayedTasksAfterShutdownPolicy(false);
-    executor.setRemoveOnCancelPolicy(true);
-
-    return executor;
-  }
-
-  private static ArchiverRepository createRepository(
-      final ExporterConfiguration config,
-      final ExporterResourceProvider resourceProvider,
-      final int partitionId,
-      final Executor executor,
-      final CamundaExporterMetrics metrics,
-      final Logger logger) {
-    final var listViewTemplate =
-        resourceProvider.getIndexTemplateDescriptor(ListViewTemplate.class);
-    final var batchOperationTemplate =
-        resourceProvider.getIndexTemplateDescriptor(BatchOperationTemplate.class);
-    return switch (ConnectionTypes.from(config.getConnect().getType())) {
-      case ELASTICSEARCH -> {
-        final var connector = new ElasticsearchConnector(config.getConnect());
-        yield new ElasticsearchRepository(
-            partitionId,
-            config.getArchiver(),
-            config.getRetention(),
-            listViewTemplate.getFullQualifiedName(),
-            batchOperationTemplate.getFullQualifiedName(),
-            connector.createAsyncClient(),
-            executor,
-            metrics,
-            logger);
-      }
-      case OPENSEARCH -> {
-        final var connector = new OpensearchConnector(config.getConnect());
-        yield new OpenSearchRepository(
-            partitionId,
-            config.getArchiver(),
-            config.getRetention(),
-            listViewTemplate.getFullQualifiedName(),
-            batchOperationTemplate.getFullQualifiedName(),
-            connector.createAsyncClient(),
-            executor,
-            metrics,
-            logger);
-      }
-    };
-  }
-
-  private record ProcessInstanceDependantAdapter(String name, String field)
-      implements ProcessInstanceDependant {
-
-    @Override
-    public String getFullQualifiedName() {
-      return name;
-    }
-
-    @Override
-    public String getProcessInstanceDependantField() {
-      return field;
     }
   }
 }
